Fix finite*() definitions in include/api/math.h
include/api/math.h included wrong declaration for finitef(), finite()
and finitel() - these should always return int (basically, a boolean),
*not* a floating-point number.

Probably before commit 4e8076829f3e8126b967f1d5b51c10b02ff290ba removed
the duplicate math.h, this didn't matter. But now it matters and
compilation fails because of the declaration is inconsistent with the
actual definition.

+++ include/api/math.h
@@ -212,9 +212,9 @@ double      fmod(double, double);
 float       fmodf(float, float);
 long double fmodl(long double, long double);
 
-double      finite(double);
-float       finitef(float);
-long double finitel(long double);
+int finite(double);
+int finitef(float);
+int finitel(long double);
 
 double      frexp(double, int *);
 float       frexpf(float, int *);
